Updated EDA notebooks
