--- bad_images.md
@@ -1,12 +1,0 @@
-![не репрезентативно](archive/sneakers-dataset/sneakers-dataset/adidas_forum_high/0142.jpg)
-![не репрезентативно](archive/sneakers-dataset/sneakers-dataset/adidas_forum_high/0146.jpg)
-![битая](archive/sneakers-dataset/sneakers-dataset/adidas_forum_low/0079.png)
-![битая](archive/sneakers-dataset/sneakers-dataset/adidas_samba/0036.png)
-![рисунок](archive/sneakers-dataset/sneakers-dataset/adidas_ultraboost/0048.jpg)
-![битая](archive/sneakers-dataset/sneakers-dataset/converse_chuck_taylor_all-star_high/0030.png)
-![битая](archive/sneakers-dataset/sneakers-dataset/converse_chuck_taylor_all-star_low/0048.png)
-![битая](archive/sneakers-dataset/sneakers-dataset/new_balance_990/0039.png)
-![битая](archive/sneakers-dataset/sneakers-dataset/nike_air_jordan_1_low/0069.png)
-![битая](archive/sneakers-dataset/sneakers-dataset/nike_dunk_low/0037.png)
-![битая](archive/sneakers-dataset/sneakers-dataset/reebok_classic_leather/0086.png)
-![битая](archive/sneakers-dataset/sneakers-dataset/vans_slip-on_checkerboard/0045.png)
