+++ notebooks/2-exploration/bad_images.md
@@ -1,0 +1,23 @@
+![не репрезентативно](../../data/01_raw/sneakers-dataset/adidas_forum_high/0142.jpg)
+
+![не репрезентативно](../../data/01_raw/sneakers-dataset/adidas_forum_high/0146.jpg)
+
+![битая](../../data/01_raw/sneakers-dataset/adidas_forum_low/0079.png)
+
+![битая](../../data/01_raw/sneakers-dataset/adidas_samba/0036.png)
+
+![рисунок](../../data/01_raw/sneakers-dataset/adidas_ultraboost/0048.jpg)
+
+![битая](../../data/01_raw/sneakers-dataset/converse_chuck_taylor_all-star_high/0030.png)
+
+![битая](../../data/01_raw/sneakers-dataset/converse_chuck_taylor_all-star_low/0048.png)
+
+![битая](../../data/01_raw/sneakers-dataset/new_balance_990/0039.png)
+
+![битая](../../data/01_raw/sneakers-dataset/nike_air_jordan_1_low/0069.png)
+
+![битая](../../data/01_raw/sneakers-dataset/nike_dunk_low/0037.png)
+
+![битая](../../data/01_raw/sneakers-dataset/reebok_classic_leather/0086.png)
+
+![битая](../../data/01_raw/sneakers-dataset/vans_slip-on_checkerboard/0045.png)
