+++ notebooks/2-exploration/eda_insights_draft.md
@@ -1,0 +1,15 @@
+## Тезисно описываю инсайты, которые вынесли благодаря EDA
+
+### Анализ [Popular Sneakers Classification](https://www.kaggle.com/datasets/nikolasgegenava/sneakers-classification?resource=download)
+
+1. Относительно низкое разрешение изображений, по сравнению с доступными в Google Images: как следствия у этого могут быть (?). Мб нужно брать исходное изображение в высоком разрешении, потом детектить на нём кроссовки, и уже обрезанную картинку сжимать до нужного (для нейронки) разрешения.
+2. yeezy_slide — не кроссовки
+3. Хотя по моделям датасет сбалансирован, видим сильный дисбаланс по брендам: почти половину выборки занимают Nike.
+4. Есть битые картинки типа такой 
+![битая картинка](../archive/sneakers-dataset/sneakers-dataset/reebok_classic_leather/0086.png).
+Чтобы избавиться от плохих изображений, 6000 изображений не сложно просмотреть: открываем Finder и записываем пути до плохих картинок в список.
+Выборочно просмотрели и поняли, что все битые изображения одинаковы: можно их идентифицировать по цветовой палитре.
+5. Граница по однородности фона — 0.04. Только 61% выборки имеет однотонный фон — довольно сбалансированно.
+6. Есть изображения, на которых не одна и не две кроссовки, а много. По хорошему их надо обрезать на несколько изображений. Пример: ![пример](../archive/sneakers-dataset/sneakers-dataset/new_balance_992/0104.jpg)
+7. На большинстве изображений одна кроссовка, «в профиль», имеют однотонный фон — именно такие изображения обычно используются на карточке товара в онлайн-магазинах.  В жизни фон будет неоднородным, ракурс будет скорее сверху вниз (т.к. кроссовки будут надеты на чьи-то ноги). В описываемом датасете присутствуют изображения, приближённые к реальной жизни, например, ![вот](../archive/sneakers-dataset/sneakers-dataset/vans_old_skool/0003.jpg)
+
