--- notebooks/eda_insights.md
@@ -1,60 +1,0 @@
-## Разведочный анализ данных
-
-### Какие данные нам нужны?
-
-В рамках проекта мы решаем задачу поиска похожих кроссовок по их изображению. Возможным решением этой задачи является следующий пайплайн: извлекаем из изображения важные для пользователя характеристики (цвет, материал изготовления, форма) и применяем их в фильтрах для поиска в онлайн-магазине.
-
-Проблемой предложенного пайплайна является то, что не все важные характеристики доступны для фильтрации: например, форму кроссовок сложно ёмко описать словами, поэтому такой фильтр обычно отсутствует в поиске онлайн-магазина (например, на Lamoda можно выбрать только между вариантами «Высокие кроссовки» и «Низкие кроссовки»).
-
-Для решения этой проблемы мы хотим воспользоваться тем, что форма кроссовки обычно определяется её моделью. Значит, чтобы найти наиболее похожие кроссовки с точки зрения формы, нам достаточно классифицировать изображение по модели — решить задачу многоклассовой классификации.
-
-В дальнейшем мы сможем совместить поиск по форме с экстракцией других характеристик (цвета, материала) и предлагать пользователю сортировать выдачу в зависимости от того, какая из характеристик представляется ему самой важной.
-
-Сейчас предлагаем сконцентрироваться на определении модели кроссовки по её изображению.
-
-### Описание данных
-
-Задачу определения модели кроссовок по их изображению уже решали, поэтому анализируем существуюший датасет, и по итогам его анализа поймём, нужно ли нам собирать дополнительные данные самостоятельно, или же этих данных достаточно.
-
-Мы провели EDA датасета [Popular Sneakers Classification](https://www.kaggle.com/datasets/nikolasgegenava/sneakers-classification?resource=download) с Kaggle.
-
-#### Размер выборки
-
-Датасет содержит 5953 изображений кроссовок, сгруппированных по модели. 145 изображений, которые относятся к модели ‘Yeezy Slide’, мы исключаем из рассмотрения, поскольку они не укладываются в наши представления о кроссовках:
-
-![yeezy_slide](../illustrations/yeezy_slide.jpg)
-
-Также мы просмотрели все изображения и отфильтровали 9 битых картинок, 1 рисунок (хотим работать с фотографиями) и 2 изображения модифицированных кроссовок. Итого осталось 5796 изображений.
-
-#### Баланс классов
-
-Итого в датасете осталось 49 моделей кроссовок. Видим дисбаланс по моделям: в самом большом классе 150 изображений, в самом маленьком 73 — разница в два раза. В среднем в одном классе 118 изображений.
-
-В датасете представлены кроссовки 10 брендов. Дисбаланс по брендам выражен существенно более сильно, чем по моделям: почти половина изображений (2293) соответствует Nike, наименее представленный бренд Asics имеет только 91 изображение.
-
-#### Разрешение изображений
-
-Среднее разрешение изображения составляет 138 на 128 пикселей (максимум по ширине — 162 пикселя). Это разрешение существенно ниже, чем имеют доступные в Google Images изображения (ширина порядка 1000 пикселей). Хотя высокое разрешение не является необходимым условием для успешного обучения модели-классификатора, нам нужно учитывать низкое разрешение в предобработке данных: рассмотреть детекцию (boxing) кроссовок на изображении, понижать разрешение фотографии, которую загружает пользователь, при инференсе модели.
-
-Возможно, получится улучшить качество модели, если собрать аналогичный датасет из фотографий высокого разрешения, применить к ним boxing, и только потом понизить разрешение для обучения нейросети (т.к. сейчас кроссовки не обязательно центрированы, они могут занимать малую долю изображения, и поэтому после обрезки от них останется совсем мало пикселей).
-
-#### Однотонность фона
-
-Мы проанализировали однотонность фона путём расчёта дисперсии цвета в 5% рамке изображения. На примерах в ноутбуке `EDA.ipynb` видно, что границу между однотоннымным и неоднотонным фоном проходит в районе метрики 0.03. В таком случае к однотонным относится 2/3 нашего датасета. Эта оценка занижена, потому что мы записали в неоднотонные также изображения, где кроссовки заходят на края изображения (т.е. где нет «рамки» из фона). 
-
-Преобладание изображений с однотонным фоном легко объяснить тем, что именно такой фон имеют стоковые изображения из онлайн-магазинов.
-
-Возможно, имеет смысл обогатить датасет изображениями, которые больше будут похожи на фотографии пользователей, чем на карточки товаров, чтобы метрика качества на обучении была более репрезентативна тому, как модель будет использоваться на проде.
-
-#### Наблюдения без статистического расчёта
-
-При просмотре изображений из датасета заметили, что особенно часто встречаются стоковые варианты: одна кроссовка изображена «в профиль» на белом фоне. Они не превалируют над другими изображениями, но для наших целей особенно важно, чтобы модель хорошо классифицировала фотографии на разнородном фоне, вид скорее сверху, чем сборку. Возможно, валидационную и тестовую выборку следует собирать именно из таких фотографий, а не делать её случайным сэмплом из датасета.
-
-В датасете также встречаются изображения более чем двух кроссовок. Для улучшения качества модели следует рассмотреть возможность разделить такие изображения на несколько.
-
-### Выводы
-
-0. Удалили мусорные изображения, которые могли только ухудшить качество итоговой модели.
-1. Датасет сильно разбалансирован по брендам, и имеет смысл дособирать изображения наименее представленных брендов и моделей.
-2. В датасете больше представлены картинки с однотонным фоном, чем с неоднотонным фоном, и это плохо, т.к. в итоге модель должна будет классифицировать фотографии пользователей с неоднотонным фоном.
-3. Сравнительно низкое разрешение изображений: возможно, получится улучшить результат, если сначала собрать картинки высокого качества, применить к ним boxing, и только потом обучать модель.
