+++ requirements.txt
@@ -7,3 +7,4 @@ scipy
 opencv-python
 matplotlib
 scikit-learn
+imagehash
